{{cookiecutter.project_slug}}/requirements/common.txt
Django==4.2.2
pytz==2019.3
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.1
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
django-guardian==2.4.0
django-redis==5.2.0
joblib==1.2.0
django-prometheus==2.3.1
drf-extensions==0.7.1
boto3==1.24.46
botocore==1.27.88
redis==4.4.2
django-storages==1.13.2
kubernetes==25.3.0
rethinkdb==2.4.9
newrelic==8.5.0
requests==2.28.2
sentry-sdk==1.14.0
beautifulsoup4==4.11.1
psycopg2-binary
mysqlclient